First proposal of macro to implement SyncPacket
