+++ ethcore/sync/src/chain/sync_packet.rs
@@ -22,6 +22,8 @@
 //! to convert to/from the packet id values transmitted over the
 //! wire.
 
+
+
 use api::{ETH_PROTOCOL, WARP_SYNC_PROTOCOL_ID};
 use network::{PacketId, ProtocolId};
 
@@ -31,86 +33,36 @@ use network::{PacketId, ProtocolId};
 /// to enum variants. This implicitly provides a mechanism to
 /// check whether a given packet id is known, and to prevent
 /// packet id clashes when defining new ids.
-enum_from_primitive! {
-#[derive(Clone, Copy, Debug, PartialEq)]
+#[derive(Clone, Copy, Debug, SyncPackets, PartialEq)]
 pub enum SyncPacket {
-	StatusPacket = 0x00,
-	NewBlockHashesPacket = 0x01,
-	TransactionsPacket = 0x02,
-	GetBlockHeadersPacket = 0x03,
-	BlockHeadersPacket = 0x04,
-	GetBlockBodiesPacket = 0x05,
-	BlockBodiesPacket = 0x06,
-	NewBlockPacket = 0x07,
+	#[eth] StatusPacket = 0x00,
+	#[eth] NewBlockHashesPacket = 0x01,
+	#[eth] TransactionsPacket = 0x02,
+	#[eth] GetBlockHeadersPacket = 0x03,
+	#[eth] BlockHeadersPacket = 0x04,
+	#[eth] GetBlockBodiesPacket = 0x05,
+	#[eth] BlockBodiesPacket = 0x06,
+	#[eth] NewBlockPacket = 0x07,
 
-	GetNodeDataPacket = 0x0d,
-	NodeDataPacket = 0x0e,
-	GetReceiptsPacket = 0x0f,
-	ReceiptsPacket = 0x10,
+	#[eth] GetNodeDataPacket = 0x0d,
+	#[eth] NodeDataPacket = 0x0e,
+	#[eth] GetReceiptsPacket = 0x0f,
+	#[eth] ReceiptsPacket = 0x10,
 
-	GetSnapshotManifestPacket = 0x11,
-	SnapshotManifestPacket = 0x12,
-	GetSnapshotDataPacket = 0x13,
-	SnapshotDataPacket = 0x14,
-	ConsensusDataPacket = 0x15,
-	PrivateTransactionPacket = 0x16,
-	SignedPrivateTransactionPacket = 0x17,
-}
-}
-
-use self::SyncPacket::*;
-
-/// Provide both subprotocol and packet id information within the
-/// same object.
-pub trait PacketInfo {
-	fn id(&self) -> PacketId;
-	fn protocol(&self) -> ProtocolId;
+	#[par] GetSnapshotManifestPacket = 0x11,
+	#[par] SnapshotManifestPacket = 0x12,
+	#[par] GetSnapshotDataPacket = 0x13,
+	#[par] SnapshotDataPacket = 0x14,
+	#[par] ConsensusDataPacket = 0x15,
+	#[par] PrivateTransactionPacket = 0x16,
+	#[par] SignedPrivateTransactionPacket = 0x17,
 }
 
-// The mechanism to match packet ids and protocol may be improved
-// through some macro magic, but for now this works.
-impl PacketInfo for SyncPacket {
-	fn protocol(&self) -> ProtocolId {
-		match self {
-			StatusPacket |
-			NewBlockHashesPacket |
-			TransactionsPacket |
-			GetBlockHeadersPacket |
-			BlockHeadersPacket |
-			GetBlockBodiesPacket |
-			BlockBodiesPacket |
-			NewBlockPacket |
-
-			GetNodeDataPacket|
-			NodeDataPacket |
-			GetReceiptsPacket |
-			ReceiptsPacket
-
-				=> ETH_PROTOCOL,
-
-			GetSnapshotManifestPacket|
-			SnapshotManifestPacket |
-			GetSnapshotDataPacket |
-			SnapshotDataPacket |
-			ConsensusDataPacket |
-			PrivateTransactionPacket |
-			SignedPrivateTransactionPacket
-
-				=> WARP_SYNC_PROTOCOL_ID,
-		}
-	}
-
-	fn id(&self) -> PacketId {
-		(*self) as PacketId
-	}
-}
 
 
 #[cfg(test)]
 mod tests {
 	use super::*;
-
-	use enum_primitive::FromPrimitive;
 
 	#[test]
 	fn packet_ids_from_u8_when_from_primitive_zero_then_equals_status_packet() {
